+++ packages/core-infrastructure/frontend/src/index.ts
@@ -1,4 +1,8 @@
-export { createActorWithConfig, loadConfig } from "./config";
+export {
+	createActorWithConfig,
+	loadConfig,
+	loadMockBackendFromModules,
+} from "./config";
 export { useActor } from "./hooks/useActor";
 export {
 	type InternetIdentityContext,
@@ -7,4 +11,9 @@ export {
 	type Status,
 	useInternetIdentity,
 } from "./hooks/useInternetIdentity";
-export type { CreateActorOptions, createActorFunction } from "./types";
+export type {
+	CreateActorOptions,
+	createActorFunction,
+	MockBackendOptions,
+	MockModules,
+} from "./types";